车型级联回填、金融计算

+++ src/main/java/com/yunche/loan/config/constant/LoanProcessConst.java
@@ -10,15 +10,15 @@ public class LoanProcessConst {
     /**
      * 打回
      */
-    public static final Byte REJECT = 0;
+    public static final Integer REJECT = 0;
     /**
      * 通过
      */
-    public static final Byte PASS = 1;
+    public static final Integer PASS = 1;
     /**
      * 弃单
      */
-    public static final Byte CANCEL = 2;
+    public static final Integer CANCEL = 2;
     /**
      * 资料增补
      */
@@ -27,19 +27,19 @@ public class LoanProcessConst {
     /**
      * 任务状态:  0-全部;
      */
-    public static final Byte TASK_ALL = 0;
+    public static final Integer TASK_ALL = 0;
     /**
      * 任务状态:  1-已处理(审核);
      */
-    public static final Byte TASK_DONE = 1;
+    public static final Integer TASK_DONE = 1;
     /**
      * 任务状态:  2-未处理(待审核);
      */
-    public static final Byte TASK_TODO = 2;
+    public static final Integer TASK_TODO = 2;
     /**
      * 任务状态：还未执行到当前任务节点
      */
-    public static final Byte TASK_NOT_REACH_CURRENT = 3;
+    public static final Integer TASK_NOT_REACH_CURRENT = 3;
 
     /**
      * 删除理由  【task被删除的理由-任务已被执行】
