+++ src/main/java/com/yunche/loan/domain/vo/AppLoanOrderVO.java
@@ -37,5 +37,5 @@ public class AppLoanOrderVO {
     /**
      * 当前任务节点审核状态:  0-全部;   1-已提交(审核);   2-未提交(审核);
      */
-    private Byte taskStatus;
+    private Integer taskStatus;
 }
