+++ src/main/java/com/yunche/loan/domain/vo/LoanOrderVO.java
@@ -47,5 +47,5 @@ public class LoanOrderVO {
     /**
      * 当前任务节点审核状态:  0-全部;   1-已提交(审核);   2-未提交(审核);
      */
-    private Byte taskStatus;
+    private Integer taskStatus;
 }
