+++ src/main/java/com/yunche/loan/domain/vo/TaskStateVO.java
@@ -13,5 +13,5 @@ public class TaskStateVO {
 
     private String taskName;
 
-    private Byte taskStatus;
+    private Integer taskStatus;
 }
